EDMT-286 refactor : 이메일 재인증 로직 auth에서 profile로 폴더 이동 및 그에 따른 리팩토링
