+++ src/hooks/api/profile/use-post-send-email.ts
@@ -1,6 +1,6 @@
 import { useMutation } from '@tanstack/react-query';
 
-import { postSendEmail } from '@/services/auth/post-send-email';
+import { postSendEmail } from '@/services/profile/post-send-email';
 import type { ApiResponseWithoutData } from '@/types/api/response';
 
 export const usePostSendEmail = () => {
